requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
fastapi==0.113.0
h11==0.14.0
idna==3.8
joblib==1.4.2
numpy==2.1.1
nvidia-nccl-cu12==2.22.3
pandas==2.2.2
pydantic==2.9.0
pydantic_core==2.23.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
